refactor: replace ConnectProjectClient with FlowsClient and remove deprecated flows module
