+++ marketplace/clients/flows/client.py
@@ -200,3 +200,153 @@ class FlowsClient(RequestClient):
             return channel
 
         return list(map(map_org_to_project_uuid, response.json()))
+
+    def list_channel_types(self, channel_code):
+        if channel_code:
+            url = f"{self.base_url}/api/v2/internals/channels/{str(channel_code)}"
+        else:
+            url = f"{self.base_url}/api/v2/internals/channels"
+
+        response = self.make_request(
+            url, method="GET", headers=self.authentication_instance.headers
+        )
+
+        return response
+
+    def detail_external(self, flow_object_uuid):
+        url = f"{self.base_url}/api/v2/internals/externals/{str(flow_object_uuid)}"
+
+        response = self.make_request(
+            url, method="GET", headers=self.authentication_instance.headers
+        )
+        return response.json()
+
+    def update_external_config(self, data, flow_object_uuid):
+        payload = {"config": data}
+        url = f"{self.base_url}/api/v2/internals/externals/{flow_object_uuid}/"
+
+        response = self.make_request(
+            url,
+            method="PATCH",
+            headers=self.authentication_instance.headers,
+            json=payload,
+        )
+        return response
+
+    def release_external_service(self, uuid: str, user_email: str):
+        url = f"{self.base_url}/api/v2/internals/externals/{uuid}/"
+        params = {"user": user_email}
+
+        response = self.make_request(
+            url,
+            method="DELETE",
+            headers=self.authentication_instance.headers,
+            params=params,
+        )
+        return response
+
+    def create_external_service(
+        self, user: str, project: str, type_fields: dict, type_code: str
+    ):
+        body = dict(
+            user=user, project=project, type_fields=type_fields, type_code=type_code
+        )
+        url = f"{self.base_url}/api/v2/internals/externals"
+
+        response = self.make_request(
+            url,
+            method="POST",
+            headers=self.authentication_instance.headers,
+            json=body,
+        )
+
+        return response
+
+    def create_prompts(self, user: str, text: str, external_uuid: str):
+        body = dict(
+            user=user,
+            text=text,
+        )
+        url = f"{self.base_url}/api/v2/internals/externals/{external_uuid}/prompts/"
+
+        response = self.make_request(
+            url,
+            method="POST",
+            headers=self.authentication_instance.headers,
+            json=body,
+        )
+        return response
+
+    def detail_prompts(self, external_uuid: str, prompt_uuid: str):
+        url = f"{self.base_url}/api/v2/internals/externals/{external_uuid}/prompts/{prompt_uuid}"
+
+        response = self.make_request(
+            url,
+            method="GET",
+            headers=self.authentication_instance.headers,
+        )
+        return response
+
+    def list_prompts(self, external_uuid: str):
+        url = f"{self.base_url}/api/v2/internals/externals/{external_uuid}/prompts/"
+
+        response = self.make_request(
+            url,
+            method="GET",
+            headers=self.authentication_instance.headers,
+        )
+        return response
+
+    def delete_prompts(self, external_uuid: str, prompt_uuid: str):
+        url = f"{self.base_url}/api/v2/internals/externals/{external_uuid}/prompts/{prompt_uuid}"
+
+        response = self.make_request(
+            url,
+            method="DELETE",
+            headers=self.authentication_instance.headers,
+        )
+        return response
+
+    def get_user_api_token(self, user: str, project_uuid: str):
+        params = dict(user=user, project=str(project_uuid))
+        url = f"{self.base_url}/api/v2/internals/users/api-token/"
+        response = self.make_request(
+            url,
+            method="GET",
+            headers=self.authentication_instance.headers,
+            params=params,
+        )
+        return response
+
+    def list_availables_channels(self):
+        url = f"{self.base_url}/v1/channel-types"
+        response = self.make_request(
+            url,
+            method="GET",
+            headers=self.authentication_instance.headers,
+        )
+        return response
+
+    def detail_channel_type(self, channel_code: str):
+        params = {"channel_type_code": channel_code}
+        url = f"{self.base_url}/v1/channel-types"
+        response = self.make_request(
+            url,
+            method="GET",
+            headers=self.authentication_instance.headers,
+            params=params,
+        )
+        return response
+
+    def release_channel(
+        self, channel_uuid: str, project_uuid: str, user_email: str
+    ) -> None:
+        params = {"user": user_email}
+        url = f"{self.base_url}/api/v2/internals/channel/{channel_uuid}"
+        self.make_request(
+            url,
+            method="DELETE",
+            headers=self.authentication_instance.headers,
+            params=params,
+        )
+        return None
