+++ marketplace/clients/router/client.py
@@ -1,0 +1,36 @@
+"""Client for WPP Router integration"""
+
+import requests
+
+from django.conf import settings
+
+
+class ConnectAuth:
+    def __get_auth_token(self) -> str:
+        request = requests.post(
+            url=settings.OIDC_OP_TOKEN_ENDPOINT,
+            data={
+                "client_id": settings.OIDC_RP_CLIENT_ID,
+                "client_secret": settings.OIDC_RP_CLIENT_SECRET,
+                "grant_type": "client_credentials",
+            },
+        )
+        token = request.json().get("access_token")
+        return f"Bearer {token}"
+
+    def auth_header(self) -> dict:
+        return {"Authorization": self.__get_auth_token()}
+
+
+class WPPRouterChannelClient(ConnectAuth):
+    base_url = settings.ROUTER_BASE_URL
+
+    def get_channel_token(self, uuid: str, name: str) -> str:
+        payload = {"uuid": uuid, "name": name}
+        response = requests.post(
+            url=self.base_url + "/integrations/channel",
+            json=payload,
+            headers=self.auth_header(),
+            timeout=60,
+        )
+        return response.json().get("token", "")
