+++ marketplace/core/types/channels/instagram/tests/test_views.py
@@ -188,9 +188,7 @@ class ConfigureInstagramAppTestCase(PermissionTestCaseMixin, APIBaseTestCase):
     def view(self):
         return self.view_class.as_view({"patch": "configure"})
 
-    @patch(
-        "marketplace.core.types.channels.instagram.views.ConnectProjectClient.create_channel"
-    )
+    @patch("marketplace.clients.flows.client.FlowsClient.create_channel")
     def test_configure_instagram_success(self, mock_create_external_service):
         mock_create_external_service.return_value = {
             "channelUuid": str(uuid.uuid4()),
@@ -209,9 +207,7 @@ class ConfigureInstagramAppTestCase(PermissionTestCaseMixin, APIBaseTestCase):
         response = self.request.patch(self.url, self.payload, uuid=self.app.uuid)
         self.assertEqual(response.status_code, 403)
 
-    @patch(
-        "marketplace.core.types.channels.instagram.views.ConnectProjectClient.create_channel"
-    )
+    @patch("marketplace.clients.flows.client.FlowsClient.create_channel")
     def test_configure_instagram_with_internal_permission_only(
         self, mock_create_external_service
     ):
