+++ marketplace/core/types/externals/chatgpt/views.py
@@ -2,7 +2,7 @@ from rest_framework.response import Response
 from rest_framework import status
 from rest_framework.decorators import action
 
-from marketplace.flows.client import FlowsClient
+from marketplace.clients.flows.client import FlowsClient
 
 from .serializers import (
     ChatGPTSerializer,
