+++ marketplace/projects/consumers/project_consumer.py
@@ -8,7 +8,8 @@ from ..usecases import (
     ProjectCreationDTO,
 )
 from marketplace.applications.usecases import AppConfigurationUseCase
-from marketplace.connect.client import ConnectProjectClient, WPPRouterChannelClient
+from marketplace.clients.flows.client import FlowsClient
+from marketplace.clients.router.client import WPPRouterChannelClient
 from marketplace.event_driven.parsers import JSONParser
 from marketplace.event_driven.consumers import EDAConsumer
 
@@ -30,11 +31,9 @@ class ProjectConsumer(EDAConsumer):  # pragma: no cover
                 authorizations=body.get("authorizations", []),
             )
 
-            connect_client = ConnectProjectClient()
+            client = FlowsClient()
             wpp_router_client = WPPRouterChannelClient()
-            app_configuration = AppConfigurationUseCase(
-                connect_client, wpp_router_client
-            )
+            app_configuration = AppConfigurationUseCase(client, wpp_router_client)
 
             app_setup_handler = AppSetupHandlerUseCase(app_configuration)
             template_type_integration = TemplateTypeIntegrationUseCase(
